Solution to Print Multiplication Table

+++ Print Multiplication Table@@676c4885c18c9c17a0f50668/code.c
@@ -1,13 +1,13 @@
 #include<stdio.h>
 
 int main() {
-    int i;
-    int n;
-    scanf("%d", &n);
+    float i;
+    float n;
+    scanf("%f", &n);
     
     for(i = n; i <= 1000; i++) {
         if(i % n == 0) {
-            printf("%d\n", i);  // Corrected: replaced /n with \n for proper newline
+            printf("%f\n", i);
         }
     }
